compile warning-- MERGE: none
svn path=/trunk/KDE/kdepim/; revision=1038670

+++ koeventview.cpp
@@ -250,12 +250,13 @@ bool KOEventView::processKeyEvent( QKeyEvent *ke )
     case Qt::Key_Alt:
       break;
     default:
-      mTypeAheadEvents.append( new QKeyEvent( ke->type(),
-                                              ke->key(),
-                                              ke->modifiers(),
-                                              ke->text(),
-                                              ke->isAutoRepeat(),
-                                              ke->count() ) );
+      mTypeAheadEvents.append(
+        new QKeyEvent( ke->type(),
+                       ke->key(),
+                       ke->modifiers(),
+                       ke->text(),
+                       ke->isAutoRepeat(),
+                       static_cast<ushort>( ke->count() ) ) );
       if ( !mTypeAhead ) {
         mTypeAhead = true;
         emit newEventSignal();
